+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Changes in version 1.x.x - 201x-xx-xx
+
+ * Medium changes
+    - Parse flag thresholds in bridge network statuses, and parse the
+      "ignoring-advertised-bws" flag threshold in relay network status
+      votes.
+
+
 # Changes in version 1.0.0 - 2015-12-05
 
  * Major changes